전남대 BE_이상현 2단계 - 페이지네이션 (#172)
* refactor: Entity 기본 생성자 접근 제한자 수정

* refactor: Product의 price 타입을 BigDeciaml로 수정

* chore: DDL 동작 쿼리를 확인하기 위해 옵션 추가

* docs: README 양식 피드백 반영

* feat: MemberRepository 테스트 코드 작성

* feat: ProductRepository 테스트 코드 작성

* docs: 기능 요구 사항 README에 작성

* refactor: 페이지네이션을 위한 Product Service, Controller 수정

* feat: 상품을 관리하는 관리자 페이지에 페이지네이션 구현

* refactor: 페이지네이션을 위한 Wish Controller, Service, Repository 수정

* refactor: 페이지네이션 반영 후 WishRepository Test 코드 수정

* refactor: 상품 페이지네이션 기본값 설정

* chore: 사용하지 않는 import문 제거

* docs: 진행 사항 README에 반영

+++ src/main/java/gift/dto/ProductResponseDto.java
@@ -1,9 +1,11 @@
 package gift.dto;
 
+import java.math.BigDecimal;
+
 public record ProductResponseDto(
     Long id,
     String name,
-    long price,
+    BigDecimal price,
     String imageUrl
 ) {
 }
